Restructuring and Logging introduction
We're now logging helo, ehlo and mails into a mail.json.
Also, the structure was updated for pluggability in the future.
Currently, RST does not work.
Also, if vrfy, mail or rcpt handlers are used,
the server crashes because somethig is not correctly set up
in the status variable.
Clear to do here.